1. rename： play.py -> cchess.py center.py -> game.py source -> src 2. updated README.md
