+++ README.md
@@ -1,5 +1,19 @@
 # Xiangqi
 
-> 基本的图形界面已完成，可打包下载查看一下效果
->
-> 运行程序 python play.py
+## 运行
+
+python cchess.py
+
+## 依赖
+
+pygame
+
+## 文件结构
+
+src: 存放棋盘、棋子、音乐文件
+
+cchess.py: 象棋游戏主体
+
+elements.py: 棋子行为
+
+game.py: 游戏基本规则
