+++ src/black_king.gif
