+++ src/black_pawn.gif
