+++ src/black_rook.gif
