+++ src/red_king.gif
