+++ src/red_pawn.gif
