+++ src/red_rook.gif
